+++ docs/source/_static/nb-scripts/3site_cycle_fin.py
@@ -14,7 +14,7 @@
 # Also plots the individual on-site energies, band structure, and Wannier
 # center of lowest band.
 
-# In[1]:
+# In[8]:
 
 
 from pythtb import TBModel, WFArray, Mesh
@@ -24,7 +24,7 @@ import matplotlib.pyplot as plt
 
 # Define function to construct model
 
-# In[2]:
+# In[9]:
 
 
 def set_model(t, delta, lmbd):
@@ -45,26 +45,32 @@ def set_model(t, delta, lmbd):
 # 
 # Next, we construct the `Mesh` object for a 1D parameter path and 1D k-space. 
 # 
+# :::{versionadded} 2.0.0
+# :::
+# 
 # When using a parameter space with the mesh, we must define a model building function as we've done above. The model parameters are defined by the arguments of the function. Some of these arguments may be fixed, while others may vary. 
 # 
 # :::{warning}
 # The argument that we plan to vary must be named appropriately in the mesh via the `axis_names` argument in the `Mesh` constructor. If the name of the model parameter we are varying does not match the corresponding axis name in the mesh, the parameter will not be varied correctly. Here we are varying the `lmbd` parameter.
 # :::
 
-# In[ ]:
+# In[10]:
 
 
 mesh = Mesh(
-    dim_k=1, dim_param=1, axis_types=["k", "param"], axis_names=["kx", "lmbd"]
+    dim_k=1, dim_lambda=1, axis_types=["k", "l"], axis_names=["kx", "lmbd"]
     )
 
 
-# We next populate the `Mesh` with grid points. To do so, we use the `build_full_grid` helper function, specifying the desired shape of the grid and whether to center it around the gamma point in k-space.
+# We next populate the `Mesh` with grid points. To do so, we use the `build_grid` helper function, specifying the desired shape of the grid and whether to center it around the gamma point in k-space.
 
-# In[4]:
+# In[11]:
 
 
-mesh.build_full_grid(shape=(31, 21), gamma_centered=True)
+mesh.build_grid(shape=(31, 21), gamma_centered=True, lambda_start=0.0, lambda_stop=1.0)
+mesh.loop_axis(0, 0)
+mesh.loop_axis(1, 1)
+print(mesh)
 
 
 # Now, we initialize the `WFArray` object with the mesh we created. To solve the tight-binding model on this mesh, we can use `solve_mesh`. 
@@ -75,7 +81,7 @@ mesh.build_full_grid(shape=(31, 21), gamma_centered=True)
 # Once again, the names matter here as well. The keys in the `fixed_params` dictionary must match the names of the arguments in the model function exactly.
 # :::
 
-# In[ ]:
+# In[12]:
 
 
 # Used for initializing the Mesh
@@ -84,17 +90,11 @@ ref_model = set_model(0,0,0)
 wfa = WFArray(ref_model, mesh)
 
 
-# In[6]:
+# In[13]:
 
 
 fixed_params = {"t": -1.3, "delta": 2.0}
 wfa.solve_mesh(set_model, fixed_params)
-
-
-# In[7]:
-
-
-wfa.shape
 
 
 # :::{versionadded} 2.0.0
@@ -102,7 +102,7 @@ wfa.shape
 # 
 # To compute the Chern numbers, we will use the `WFArray.chern_num` method, which calculates the integrated Berry flux for a given set of bands and a specified plane in the Brillouin zone.
 
-# In[8]:
+# In[14]:
 
 
 # compute integrated curvature
@@ -127,7 +127,7 @@ print(f"  Band  2 = {chern_2:5.2f}")
 
 # Here, we will define a new model function for the finite system. This function will take the model parameters as input and return the corresponding `TBModel`. In this case, it will be the same as before except we cut out a finite chain of the periodic model. We do this using the `TBModel.cut_piece` method, passing the number of unit cells and the direction in which to cut.
 
-# In[9]:
+# In[15]:
 
 
 # length of chain, in unit cells
@@ -148,16 +148,17 @@ def finite_model_builder(t, delta, lmbd):
 # 
 # In the `shape` argument of `build_full_grid`, we only need to specify the size of the array of `lmbd` points. 
 
-# In[11]:
+# In[20]:
 
 
-mesh = Mesh(dim_k=0, dim_param=1, axis_types=["param"], axis_names=["lmbd"])
-mesh.build_full_grid(shape=(241,))
+mesh = Mesh(dim_k=0, dim_lambda=1, axis_types=["l"], axis_names=["lmbd"])
+mesh.build_grid(shape=(241,), lambda_start=0.0, lambda_stop=1.0)
+print(mesh)
 
 
 # Same as before, we create the `WFArray` to store our states with the mesh, and use `solve_mesh` to populate the `WFArray` with the wave functions and energies.
 
-# In[13]:
+# In[21]:
 
 
 ref_model = finite_model_builder(0, 0, 0)
@@ -170,7 +171,7 @@ wfa.solve_mesh(model_func=finite_model_builder, fixed_params=fixed_params)
 # 
 # Getting the expectation value of the position operator is as simple as calling `WFArray.position_expectation(...)`. We only need to pass the real-space direction to compute the expectation value. This will return the expectation value for each state index, at each point in the mesh array.
 
-# In[14]:
+# In[22]:
 
 
 x_expec = wfa.position_expectation(dir=0)
@@ -180,7 +181,7 @@ x_expec = wfa.position_expectation(dir=0)
 # 
 # We will indicate the position expectation value by the size of the markers in the plot. Notice in the finite chain the appearance of gapless edge modes. This is topologically protected, as evident by the Chern numbers we calculated earlier in the bulk case. 
 
-# In[15]:
+# In[23]:
 
 
 fig, ax = plt.subplots()