+++ docs/source/_static/nb-scripts/graphene_cone.py
@@ -14,7 +14,7 @@
 # Brillouin zone containing the Dirac point, and plots individual phases
 # for each plaquette in the array.
 
-# In[13]:
+# In[1]:
 
 
 from pythtb import TBModel, WFArray, Mesh
@@ -24,7 +24,7 @@ import matplotlib.pyplot as plt
 
 # First we build the tight-binding model for graphene with a staggered onsite potential.
 
-# In[14]:
+# In[2]:
 
 
 # define lattice vectors
@@ -53,7 +53,7 @@ print(my_model)
 # 
 # First we will construct the circular path of k-points around the Dirac cone.
 
-# In[15]:
+# In[3]:
 
 
 circ_step = 31 # number of steps in the circular path
@@ -78,26 +78,27 @@ for i in range(circ_step):
 # `n_interp` is an optional parameter that specifies the number of interpolation points between each pair of nodes in the path. By default, it is set to 1. This default behavior means that the k-point path will consist only of the nodes specified in `path_k`, without any additional points in between. If you want to create a denser k-point path with more points between the nodes, you can increase the value of `n_interp`.
 # :::
 
-# In[16]:
+# In[4]:
 
 
-mesh = Mesh(my_model, axis_types=['k'])
+mesh = Mesh(dim_k=2, axis_types=['k'])
 mesh.build_path(nodes=kpts, n_interp=1)
+print(mesh)
 
 
 # ### `WFArray` class
 # 
 # We now construct a `WFArray` object to hold the wavefunction data for each k-point in the mesh. The `WFArray` class is designed to work seamlessly with the `Mesh` class, allowing us to easily associate wavefunction data with the specific k-points (or parameter points) stored in the `Mesh`. 
 
-# In[17]:
+# In[5]:
 
 
-w_circ = WFArray(mesh)
+w_circ = WFArray(my_model, mesh)
 
 
-# To populate the `WFArray` object with wavefunction data, we can use the `solve_k_mesh()` method, which computes the wavefunctions for each k-point in the mesh.
+# To populate the `WFArray` object with wavefunction data, we can use the `solve_mesh()` method, which computes the wavefunctions for each k-point in the mesh.
 
-# In[18]:
+# In[6]:
 
 
 w_circ.solve_mesh()
@@ -106,12 +107,12 @@ w_circ.solve_mesh()
 # ### Berry phase
 # We can compute the Berry phase along the circular path using the `berry_phase` method of the `WFArray` object. This method takes a list of band indices as input and returns the Berry phase for those bands.
 
-# In[19]:
+# In[7]:
 
 
-berry_phase_0 = w_circ.berry_phase([0])
-berry_phase_1 = w_circ.berry_phase([1])
-berry_phase_both = w_circ.berry_phase([0, 1])
+berry_phase_0 = w_circ.berry_phase(0, [0])
+berry_phase_1 = w_circ.berry_phase(0, [1])
+berry_phase_both = w_circ.berry_phase(0, [0, 1])
 
 print(f"Berry phase along circle with radius {circ_radius} and centered at k-point {circ_center}")
 print(f"for band 0 equals    : {berry_phase_0}")
@@ -121,14 +122,14 @@ print(f"for both bands equals: {berry_phase_both}")
 
 # ## Square patch around Dirac cone
 # 
-# Next, we construct a two-dimensional square patch covering the Dirac cone.
+# Next, we construct a two-dimensional square patch covering the Dirac cone. We will construct the side length of the square patch such that the area of the patch equals the area enclosed by the loop around the Dirac point with radius `circ_radius` constructed above (`square_length` = $\sqrt{\pi \texttt{circ\_radius}^2}$)
 
-# In[20]:
+# In[8]:
 
 
-square_step = 31
+square_step = 50
 square_center = np.array([1/3, 2/3])
-square_length = 0.1
+square_length = np.sqrt(np.pi * circ_radius**2)
 
 all_kpt = np.zeros((square_step, square_step, 2))
 for i in range(square_step):
@@ -154,22 +155,21 @@ for i in range(square_step):
 # The `points` array must have a shape that corresponds to `shape_k`. For example, if `shape_k` is `(4, 4)`, then `points` should have the shape `(4, 4, 2)` to represent the k-point coordinates in 2D.
 # :::
 
-# In[21]:
+# In[9]:
 
 
-mesh = Mesh(my_model, axis_types=['k', 'k'])
-mesh.build_custom(points=all_kpt) 
-
-# mesh.build_grid(shape_k=(square_step, square_step), points=all_kpt)
+mesh = Mesh(dim_k=2, axis_types=['k', 'k'])
+mesh.build_custom(points=all_kpt)
+print(mesh) 
 
 
 # ### `WFArray` class
 # Now we do the same thing as before to solve the model on these k-points, by calling `solve_k_mesh` on the `WFArray` object.
 
-# In[22]:
+# In[10]:
 
 
-w_square = WFArray(mesh)
+w_square = WFArray(my_model, mesh)
 w_square.solve_mesh()
 
 
@@ -182,7 +182,7 @@ w_square.solve_mesh()
 # However, if `plane` is unspecified, the Berry flux will be computed for all available planes, and will be returned with an additional set of two axes corresponding to each dimension in parameter space. Since the Berry flux is an anti-symmetric tensor, the `[0,1]` and `[1,0]` components will be related by a minus sign. So here, we specify the plane so the returned object just gets the upper left (`[0,1]`) component.
 # :::
 
-# In[23]:
+# In[11]:
 
 
 b_flux_0 = w_square.berry_flux([0], plane=(0, 1))
@@ -195,9 +195,9 @@ print("for band 1 equals    : ", np.sum(b_flux_1))
 print("for both bands equals: ", np.sum(b_flux_both))
 
 
-# Let's plot the Berry flux distribution in the kx-ky plane.
+# Let's plot the Berry curvature distribution in the kx-ky plane. 
 
-# In[24]:
+# In[12]:
 
 
 fig, ax = plt.subplots()
@@ -211,9 +211,9 @@ img = ax.imshow(
         all_kpt[0, -2, 1],
     ),
 )
-ax.set_title("Berry curvature near Dirac cone")
-ax.set_xlabel(r"$k_x$")
-ax.set_ylabel(r"$k_y$")
+ax.set_title("Berry curvature of lower band near Dirac cone")
+ax.set_xlabel(r"$k_1$")
+ax.set_ylabel(r"$k_2$")
 plt.colorbar(img)
 fig.tight_layout()